api/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30
asyncpg>=0.30
pydantic-settings>=2.0
strawberry-graphql[fastapi]>=0.250
httpx>=0.27
pytest>=8.0
pytest-asyncio>=0.24
